+++ app/helpers/rails_agent_annotator/annotator_helper.rb
@@ -9,7 +9,8 @@ module RailsAgentAnnotator
     def annotator_context_payload
       context = RailsAgentAnnotator.context_for(controller).merge(
         storage_key_prefix: RailsAgentAnnotator.configuration.storage_key_prefix,
-        mount_path: RailsAgentAnnotator.configuration.mount_path
+        mount_path: RailsAgentAnnotator.configuration.mount_path,
+        app_id: RailsAgentAnnotator.resolved_app_id
       )
 
       ERB::Util.json_escape(context.to_json)