+++ lib/rails_agent_annotator/configuration.rb
@@ -1,7 +1,7 @@
 module RailsAgentAnnotator
   class Configuration
     attr_accessor :enabled_environments, :enabled, :allow_in_staging, :authorize,
-      :mount_path, :storage_key_prefix, :markdown_template
+      :mount_path, :storage_key_prefix, :markdown_template, :app_id
 
     def initialize
       @enabled_environments = %w[development]
@@ -11,6 +11,7 @@ module RailsAgentAnnotator
       @mount_path = "/__agent_annotator"
       @storage_key_prefix = "rails_agent_annotator"
       @markdown_template = :v1
+      @app_id = nil
     end
   end
 end